+++ packages/adapter-cloudflare/index.js
@@ -1,8 +1,9 @@
-import { copyFileSync, existsSync, readFileSync, writeFileSync } from 'node:fs';
 import path from 'node:path';
+import process from 'node:process';
+import { copyFileSync, existsSync, readFileSync, writeFileSync } from 'node:fs';
 import { fileURLToPath } from 'node:url';
-import { getPlatformProxy } from 'wrangler';
-import { get_routes_json, parse_redirects } from 'utils.js';
+import { get_routes_json, parse_redirects } from './utils.js';
+import { getPlatformProxy, unstable_readConfig } from 'wrangler';
 
 /** @type {import('./index.js').default} */
 export default function (options = {}) {
@@ -14,7 +15,7 @@ export default function (options = {}) {
 				existsSync(`${builder.config.kit.files.assets}/_routes.json`)
 			) {
 				throw new Error(
-					"Cloudflare's _routes.json should be configured in svelte.config.js. See https://svelte.dev/docs/kit/adapter-cloudflare#Options-routes"
+					"Cloudflare Pages' _routes.json should be configured in svelte.config.js. See https://svelte.dev/docs/kit/adapter-cloudflare#Options-routes"
 				);
 			}
 
@@ -30,77 +31,123 @@ export default function (options = {}) {
 				);
 			}
 
+			const wrangler_config = validate_config(options.config);
+			const building_for_cloudflare_pages = is_building_for_cloudflare_pages(wrangler_config);
+
+			let dest = builder.getBuildDirectory('cloudflare');
+			let worker_dest = `${dest}/_worker.js`;
+			let assets_binding = 'ASSETS';
+
+			if (building_for_cloudflare_pages) {
+				if (wrangler_config.pages_build_output_dir) {
+					dest = wrangler_config.pages_build_output_dir;
+					worker_dest = `${dest}/_worker.js`;
+				}
+			} else {
+				if (wrangler_config.main) {
+					worker_dest = wrangler_config.main;
+				}
+				if (wrangler_config.assets?.directory) {
+					dest = wrangler_config.assets.directory;
+				}
+				if (wrangler_config.assets?.binding) {
+					assets_binding = wrangler_config.assets.binding;
+				}
+			}
+
 			const files = fileURLToPath(new URL('./files', import.meta.url).href);
-			const dest = builder.getBuildDirectory('cloudflare');
 			const tmp = builder.getBuildDirectory('cloudflare-tmp');
 
 			builder.rimraf(dest);
+			builder.rimraf(worker_dest);
 
 			builder.mkdirp(dest);
 			builder.mkdirp(tmp);
 
-			// generate plaintext 404.html first which can then be overridden by prerendering, if the user defined such a page.
-			// This file is served when a request that matches an entry in `routes.exclude` fails to match an asset.
-			const fallback = path.join(dest, '404.html');
-			if (options.fallback === 'spa') {
-				await builder.generateFallback(fallback);
-			} else {
-				writeFileSync(fallback, 'Not Found');
+			// client assets and prerendered pages
+			const assets_dest = `${dest}${builder.config.kit.paths.base}`;
+			builder.mkdirp(assets_dest);
+			if (
+				building_for_cloudflare_pages ||
+				wrangler_config.assets?.not_found_handling === '404-page'
+			) {
+				// generate plaintext 404.html first which can then be overridden by prerendering, if the user defined such a page.
+				// This file is served when a request fails to match an asset.
+				// If we're building for Cloudflare Pages, it's only served when a request matches an entry in `routes.exclude`
+				const fallback = path.join(assets_dest, '404.html');
+				if (options.fallback === 'spa') {
+					await builder.generateFallback(fallback);
+				} else {
+					writeFileSync(fallback, 'Not Found');
+				}
+			}
+			const client_assets = builder.writeClient(assets_dest);
+			builder.writePrerendered(assets_dest);
+			if (
+				!building_for_cloudflare_pages &&
+				wrangler_config.assets?.not_found_handling === 'single-page-application'
+			) {
+				await builder.generateFallback(path.join(assets_dest, 'index.html'));
 			}
 
-			// client assets and prerendered pages
-			const dest_dir = `${dest}${builder.config.kit.paths.base}`;
-			const client_assets = builder.writeClient(dest_dir);
-			builder.writePrerendered(dest_dir);
-
-			const relativePath = path.posix.relative(dest, builder.getServerDirectory());
+			// worker
+			const worker_dest_dir = path.dirname(worker_dest);
 			writeFileSync(
 				`${tmp}/manifest.js`,
-				`export const manifest = ${builder.generateManifest({ relativePath })};\n\n` +
+				`export const manifest = ${builder.generateManifest({ relativePath: path.posix.relative(tmp, builder.getServerDirectory()) })};\n\n` +
 					`export const prerendered = new Set(${JSON.stringify(builder.prerendered.paths)});\n\n` +
 					`export const base_path = ${JSON.stringify(builder.config.kit.paths.base)};\n`
 			);
-			builder.copy(`${files}/worker.js`, `${dest}/_worker.js`, {
+			builder.copy(`${files}/worker.js`, worker_dest, {
 				replace: {
-					SERVER: `${relativePath}/index.js`,
-					MANIFEST: `${path.posix.relative(dest, tmp)}/manifest.js`
+					// the paths returned by the Wrangler config might be Windows paths,
+					// so we need to convert them to POSIX paths or else the backslashes
+					// will be interpreted as escape characters and create an incorrect import path
+					SERVER: `${posixify(path.relative(worker_dest_dir, builder.getServerDirectory()))}/index.js`,
+					MANIFEST: `${posixify(path.relative(worker_dest_dir, tmp))}/manifest.js`,
+					ASSETS: assets_binding
 				}
 			});
 
 			// _headers
-			if (existsSync('_headers')) {
-				copyFileSync('_headers', `${dest}/_headers`);
+			const headers_src = '_headers';
+			const headers_dest = `${dest}/_headers`;
+			if (existsSync(headers_src)) {
+				copyFileSync(headers_src, headers_dest);
 			}
-			writeFileSync(`${dest}/_headers`, generate_headers(builder.getAppPath()), { flag: 'a' });
+			writeFileSync(headers_dest, generate_headers(builder.getAppPath()), { flag: 'a' });
 
 			// _redirects
-			const redirects_file = `${dest}/_redirects`;
-			if (existsSync('_redirects')) {
-				copyFileSync('_redirects', redirects_file);
+			const redirects_src = '_redirects';
+			const redirects_dest = `${dest}/_redirects`;
+			if (existsSync(redirects_src)) {
+				copyFileSync(redirects_src, redirects_dest);
 			}
 			if (builder.prerendered.redirects.size > 0) {
-				writeFileSync(redirects_file, generate_redirects(builder.prerendered.redirects), {
+				writeFileSync(redirects_dest, generate_redirects(builder.prerendered.redirects), {
 					flag: 'a'
 				});
 			}
 
-			// _routes.json
-			/** @type {string[]} */
-			let redirects = [];
-			if (existsSync(redirects_file)) {
-				const redirect_rules = readFileSync(redirects_file, 'utf8');
-				redirects = parse_redirects(redirect_rules);
+			if (building_for_cloudflare_pages) {
+				// _routes.json
+				/** @type {string[]} */
+				let redirects = [];
+				if (existsSync(redirects_dest)) {
+					const redirect_rules = readFileSync(redirects_dest, 'utf8');
+					redirects = parse_redirects(redirect_rules);
+				}
+				writeFileSync(
+					`${dest}/_routes.json`,
+					JSON.stringify(
+						get_routes_json(builder, client_assets, redirects, options.routes ?? {}),
+						null,
+						'\t'
+					)
+				);
+			} else {
+				writeFileSync(`${dest}/.assetsignore`, generate_assetsignore(), { flag: 'a' });
 			}
-			writeFileSync(
-				`${dest}/_routes.json`,
-				JSON.stringify(
-					get_routes_json(builder, client_assets, redirects, options.routes ?? {}),
-					null,
-					'\t'
-				)
-			);
-
-			writeFileSync(`${dest}/.assetsignore`, generate_assetsignore(), { flag: 'a' });
 		},
 		emulate() {
 			// we want to invoke `getPlatformProxy` only once, but await it only when it is accessed.
@@ -184,4 +231,59 @@ _routes.json
 _headers
 _redirects
 `.trimEnd();
+}
+
+/**
+ * @param {string} config_file
+ * @returns {import('wrangler').Unstable_Config}
+ */
+function validate_config(config_file = undefined) {
+	const wrangler_config = unstable_readConfig({ config: config_file });
+
+	// we don't support workers sites
+	if (wrangler_config.site) {
+		throw new Error(
+			`You must remove all \`site\` keys in ${wrangler_config.configPath}. Consult https://svelte.dev/docs/kit/adapter-cloudflare#Migrating-from-Workers-Sites-to-Workers-Static-Assets`
+		);
+	}
+
+	if (is_building_for_cloudflare_pages(wrangler_config)) {
+		return wrangler_config;
+	}
+
+	// probably deploying to Cloudflare Workers
+	if (wrangler_config.main || wrangler_config.assets) {
+		if (!wrangler_config.assets?.directory) {
+			throw new Error(
+				`You must specify the \`assets.directory\` key in ${wrangler_config.configPath}. Consult https://developers.cloudflare.com/workers/static-assets/binding/#directory`
+			);
+		}
+
+		if (!wrangler_config.assets?.binding) {
+			throw new Error(
+				`You must specify the \`assets.binding\` key in ${wrangler_config.configPath}. Consult https://developers.cloudflare.com/workers/static-assets/binding/#binding`
+			);
+		}
+	}
+
+	return wrangler_config;
+}
+
+/**
+ * @param {import('wrangler').Unstable_Config} wrangler_config
+ * @returns {boolean}
+ */
+function is_building_for_cloudflare_pages(wrangler_config) {
+	return (
+		!!process.env.CF_PAGES ||
+		!wrangler_config.configPath ||
+		!!wrangler_config.pages_build_output_dir ||
+		!wrangler_config.main ||
+		!wrangler_config.assets
+	);
+}
+
+/** @param {string} str */
+function posixify(str) {
+	return str.replace(/\\/g, '/');
 }