+++ packages/kit/src/runtime/control.js
@@ -61,27 +61,3 @@ export class ActionFailure {
 		this.data = data;
 	}
 }
-
-/**
- * This is a grotesque hack that, in dev, allows us to replace the implementations
- * of these classes that you'd get by importing them from `@sveltejs/kit` with the
- * ones that are imported via Vite and loaded internally, so that instanceof
- * checks work even though SvelteKit imports this module via Vite and consumers
- * import it via Node
- * @param {{
- *   ActionFailure: typeof ActionFailure;
- *   HttpError: typeof HttpError;
- *   Redirect: typeof Redirect;
- *   SvelteKitError: typeof SvelteKitError;
- * }} implementations
- */
-export function replace_implementations(implementations) {
-	// @ts-expect-error
-	ActionFailure = implementations.ActionFailure; // eslint-disable-line no-class-assign
-	// @ts-expect-error
-	HttpError = implementations.HttpError; // eslint-disable-line no-class-assign
-	// @ts-expect-error
-	Redirect = implementations.Redirect; // eslint-disable-line no-class-assign
-	// @ts-expect-error
-	SvelteKitError = implementations.SvelteKitError; // eslint-disable-line no-class-assign
-}